requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
cryptography
black
pydantic[email]
fastapi[security]
python-dotenv
python-multipart
requests
python-jose
passlib
passlib[bcrypt]
bcrypt
fpdf

pillow
PyMuPDF



